requirements.txt
pyyaml>=6.0
flask>=2.0.0
werkzeug>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
cryptography>=36.0.0